refactor(streamlit): modulariza app em core/pages/visuals e adiciona CSS global
